fix: 修复价格颜色逻辑中的类型转换，使用 parseFloat 替代 parseInt

+++ src/view/CandleLastPriceLineView.ts
@@ -48,7 +48,7 @@ export default class CandleLastPriceView extends View {
             ),
             precision.price
           )
-          color = Number.parseInt(text) > 0 ? lastPriceMarkStyles.upColor : lastPriceMarkStyles.downColor
+          color = Number.parseFloat(text) > 0 ? lastPriceMarkStyles.upColor : lastPriceMarkStyles.downColor
         } else {
           if (close > comparePrice) {
             color = lastPriceMarkStyles.upColor
